improve AccumulatorArgs
