+++ datafusion/expr/src/lib.rs
@@ -46,10 +46,12 @@ pub mod expr;
 pub mod expr_fn;
 pub mod expr_rewriter;
 pub mod expr_schema;
+pub mod expressions;
 pub mod function;
 pub mod groups_accumulator;
 pub mod interval_arithmetic;
 pub mod logical_plan;
+pub mod physical_expr;
 pub mod planner;
 pub mod registry;
 pub mod simplify;
