+++ datafusion/physical-expr-common/src/aggregate/mod.rs
@@ -18,9 +18,9 @@
 use std::fmt::Debug;
 use std::{any::Any, sync::Arc};
 
-use arrow::datatypes::{DataType, Field, Schema, SchemaRef};
+use arrow::datatypes::{DataType, Field, Schema};
 
-use datafusion_common::exec_err;
+use datafusion_common::{exec_err, ToDFSchema};
 use datafusion_common::{internal_err, not_impl_err, DFSchema, Result};
 use datafusion_expr::function::StateFieldsArgs;
 use datafusion_expr::type_coercion::aggregates::check_arg_count;
@@ -30,9 +30,9 @@ use datafusion_expr::{
     function::AccumulatorArgs, Accumulator, AggregateUDF, Expr, GroupsAccumulator,
 };
 
-use crate::physical_expr::PhysicalExpr;
 use crate::sort_expr::{LexOrdering, PhysicalSortExpr};
 use crate::utils::reverse_order_bys;
+use datafusion_expr::physical_expr::PhysicalExpr;
 
 use self::utils::down_cast_any_ref;
 
@@ -76,7 +76,7 @@ pub fn create_aggregate_expr(
     builder = builder.sort_exprs(sort_exprs.to_vec());
     builder = builder.order_by(ordering_req.to_vec());
     builder = builder.logical_exprs(input_exprs.to_vec());
-    builder = builder.schema(Arc::new(schema.clone()));
+    builder = builder.dfschema(Arc::new(schema.clone()).to_dfschema()?);
     builder = builder.name(name);
 
     if ignore_nulls {
@@ -109,8 +109,6 @@ pub fn create_aggregate_expr_with_dfschema(
     builder = builder.order_by(ordering_req.to_vec());
     builder = builder.logical_exprs(input_exprs.to_vec());
     builder = builder.dfschema(dfschema.clone());
-    let schema: Schema = dfschema.into();
-    builder = builder.schema(Arc::new(schema));
     builder = builder.name(name);
 
     if ignore_nulls {
@@ -138,8 +136,6 @@ pub struct AggregateExprBuilder {
     /// Logical expressions of the aggregate function, it will be deprecated in <https://github.com/apache/datafusion/issues/11359>
     logical_args: Vec<Expr>,
     name: String,
-    /// Arrow Schema for the aggregate function
-    schema: SchemaRef,
     /// Datafusion Schema for the aggregate function
     dfschema: DFSchema,
     /// The logical order by expressions, it will be deprecated in <https://github.com/apache/datafusion/issues/11359>
@@ -161,7 +157,6 @@ impl AggregateExprBuilder {
             args,
             logical_args: vec![],
             name: String::new(),
-            schema: Arc::new(Schema::empty()),
             dfschema: DFSchema::empty(),
             sort_exprs: vec![],
             ordering_req: vec![],
@@ -177,7 +172,6 @@ impl AggregateExprBuilder {
             args,
             logical_args,
             name,
-            schema,
             dfschema,
             sort_exprs,
             ordering_req,
@@ -195,7 +189,7 @@ impl AggregateExprBuilder {
         if !ordering_req.is_empty() {
             let ordering_types = ordering_req
                 .iter()
-                .map(|e| e.expr.data_type(&schema))
+                .map(|e| e.expr.data_type(dfschema.as_arrow()))
                 .collect::<Result<Vec<_>>>()?;
 
             ordering_fields = utils::ordering_fields(&ordering_req, &ordering_types);
@@ -203,7 +197,7 @@ impl AggregateExprBuilder {
 
         let input_exprs_types = args
             .iter()
-            .map(|arg| arg.data_type(&schema))
+            .map(|arg| arg.data_type(dfschema.as_arrow()))
             .collect::<Result<Vec<_>>>()?;
 
         check_arg_count(
@@ -233,11 +227,6 @@ impl AggregateExprBuilder {
 
     pub fn name(mut self, name: impl Into<String>) -> Self {
         self.name = name.into();
-        self
-    }
-
-    pub fn schema(mut self, schema: SchemaRef) -> Self {
-        self.schema = schema;
         self
     }
 
@@ -524,8 +513,7 @@ impl AggregateExpr for AggregateFunctionExpr {
             ignore_nulls: self.ignore_nulls,
             sort_exprs: &self.sort_exprs,
             is_distinct: self.is_distinct,
-            input_types: &self.input_types,
-            input_exprs: &self.logical_args,
+            input_exprs: &self.args,
             name: &self.name,
             is_reversed: self.is_reversed,
         };
@@ -540,8 +528,7 @@ impl AggregateExpr for AggregateFunctionExpr {
             ignore_nulls: self.ignore_nulls,
             sort_exprs: &self.sort_exprs,
             is_distinct: self.is_distinct,
-            input_types: &self.input_types,
-            input_exprs: &self.logical_args,
+            input_exprs: &self.args,
             name: &self.name,
             is_reversed: self.is_reversed,
         };
@@ -611,8 +598,7 @@ impl AggregateExpr for AggregateFunctionExpr {
             ignore_nulls: self.ignore_nulls,
             sort_exprs: &self.sort_exprs,
             is_distinct: self.is_distinct,
-            input_types: &self.input_types,
-            input_exprs: &self.logical_args,
+            input_exprs: &self.args,
             name: &self.name,
             is_reversed: self.is_reversed,
         };
@@ -626,8 +612,7 @@ impl AggregateExpr for AggregateFunctionExpr {
             ignore_nulls: self.ignore_nulls,
             sort_exprs: &self.sort_exprs,
             is_distinct: self.is_distinct,
-            input_types: &self.input_types,
-            input_exprs: &self.logical_args,
+            input_exprs: &self.args,
             name: &self.name,
             is_reversed: self.is_reversed,
         };
