+++ datafusion/physical-expr-common/src/expressions/mod.rs
@@ -16,7 +16,6 @@
 // under the License.
 
 mod cast;
-pub mod column;
 pub mod literal;
 
 pub use cast::{cast, cast_with_options, CastExpr};
