+++ datafusion/physical-expr/benches/is_null.rs
@@ -20,9 +20,9 @@ use arrow::record_batch::RecordBatch;
 use arrow_array::builder::Int32Builder;
 use arrow_schema::DataType;
 use criterion::{black_box, criterion_group, criterion_main, Criterion};
+use datafusion_expr::expressions::column::Column;
+use datafusion_expr::physical_expr::PhysicalExpr;
 use datafusion_physical_expr::expressions::{IsNotNullExpr, IsNullExpr};
-use datafusion_physical_expr_common::expressions::column::Column;
-use datafusion_physical_expr_common::physical_expr::PhysicalExpr;
 use std::sync::Arc;
 
 fn criterion_benchmark(c: &mut Criterion) {
