+++ datafusion/physical-expr/src/physical_expr.rs
@@ -17,10 +17,10 @@
 
 use std::sync::Arc;
 
-use datafusion_physical_expr_common::physical_expr::PhysicalExpr;
+use datafusion_expr::physical_expr::PhysicalExpr;
 use itertools::izip;
 
-pub use datafusion_physical_expr_common::physical_expr::down_cast_any_ref;
+pub use datafusion_expr::physical_expr::down_cast_any_ref;
 
 /// Shared [`PhysicalExpr`].
 pub type PhysicalExprRef = Arc<dyn PhysicalExpr>;
